+++ client/src/main.tsx
@@ -15,7 +15,7 @@ import { PrimeReactProvider } from "primereact/api";
 import "primereact/resources/primereact.min.css";
 import "primeicons/primeicons.css";
 import "primereact/resources/themes/bootstrap4-dark-blue/theme.css";
-import { LanguageProvider } from "./contexts/LanguageContext";
+import { LanguageProvider } from "./context/LanguageContext";
 
 const queryClient = new QueryClient({
   defaultOptions: {
